Dashboard/requirements.txt
Flask
requests
flask-cors
PyYAML
